Moves ghwc client code into cmd/ghwc/ directory
`ghw` was started before the Golang community standardized best
practices of having a `cmd/` directory contain all CLI tools for a
project. This patch simply moves the location of the sample `ghwc` CLI
tool to this best practice `cmd/ghwc/` directory.

You can install ghwc using `go get github.com/jaypipes/ghw/cmd/ghwc`

Issue #90
