+++ README.md
@@ -804,7 +804,7 @@ You can run unit tests easily using the `make test` command, like so:
 
 ```
 [jaypipes@uberbox ghw]$ make test
-go test github.com/jaypipes/ghw github.com/jaypipes/ghw/ghwc
+go test github.com/jaypipes/ghw github.com/jaypipes/ghw/cmd/ghwc
 ok      github.com/jaypipes/ghw 0.084s
-?       github.com/jaypipes/ghw/ghwc    [no test files]
+?       github.com/jaypipes/ghw/cmd/ghwc    [no test files]
 ```
